requirements.txt
python-telegram-bot==20.7
langchain==0.2.0
langchain-community==0.2.0